fix(serial) make extremely fast serial input work  (#15)
* feat(examples/serialport-test) fix examples styles

* feat(examples/serialport-test) alphabetically sort ports

* fix(serial) make extremely fast serial input work

Serial connections that rapidly send a lot of data tend to have random sections of data missing. I've narrowed this down to the timeout option.

Removing the timeout altogether causes tauri apps to crash, so instead I use the timeout option as the time to send the combined buffer. This fixes that issue that occurs when too much data is sent too fast from the device.